requirements.txt
aiogram==3.10.0
aiohttp
gspread
google-auth
